refactor: improve aliases
