+++ src/components/index.ts
@@ -1,3 +1,5 @@
 export { RestfulView } from './restful/restful-view';
+export { AddButton } from './variable/add_var_button';
 export { CreateVariable } from './variable/create_variable';
 export { VariableItem } from './variable/variable_item';
+export { VariableList } from './variable/variable_list';
